refactor: move chunking logic to utils
